unbreak EOPA tests
