+++ build.gradle
@@ -124,9 +124,9 @@ if (!project.hasProperty('skip.signing')) {
 dependencies {
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
     runtimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
-    testImplementation 'org.testcontainers:testcontainers-bom:1.19.7'
-    testImplementation 'org.testcontainers:testcontainers:1.19.7'
-    testImplementation 'org.testcontainers:junit-jupiter:1.19.7'
+    testImplementation 'org.testcontainers:testcontainers-bom:1.20.1'
+    testImplementation 'org.testcontainers:testcontainers:1.20.1'
+    testImplementation 'org.testcontainers:junit-jupiter:1.20.1'
     implementation 'com.fasterxml.jackson.core:jackson-annotations:2.17.0'
     api 'com.fasterxml.jackson.core:jackson-core:2.17.0'
     implementation 'org.apache.httpcomponents:httpcore:4.4.16'
