+++ src/test/java/com/styra/opa/OPATest.java
@@ -32,7 +32,6 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Testcontainers
-
 class OPATest {
 
     private int opaPort = 8181;
@@ -48,8 +47,6 @@ class OPATest {
 
     private String address;
     private String altAddress;
-    private String eopaAddress;
-    private String eopaAltAddress;
     private Map<String, String> headers = Map.ofEntries(entry("Authorization", "Bearer supersecret"));
 
     private RecordingHandler recordingHandler;
@@ -73,29 +70,13 @@ class OPATest {
         )
         .withExposedPorts(opaPort, altPort)
         .withFileSystemBind("./testdata/simple", "/policy", BindMode.READ_ONLY)
-        .withCommand("run -s --authentication=token --authorization=basic --bundle /policy");
-    //CHECKSTYLE:ON
-
-    @Container
-    //CHECKSTYLE:OFF
-    public GenericContainer<?> eopac = new GenericContainer<>(
-            new ImageFromDockerfile()
-                .withFileFromClasspath("Dockerfile", "eopa.Dockerfile")
-                .withFileFromClasspath("nginx.conf", "nginx.conf")
-                .withFileFromClasspath("entrypoint.sh", "entrypoint.sh")
-        )
-        .withExposedPorts(opaPort, altPort)
-        .withFileSystemBind("./testdata/simple", "/policy", BindMode.READ_ONLY)
-        .withCommand("run -s --authentication=token --authorization=basic --bundle /policy")
-        .withEnv("EOPA_LICENSE_KEY", System.getenv("EOPA_LICENSE_KEY"));
+        .withCommand("run -s -a 0.0.0.0:8181 --authentication=token --authorization=basic --bundle /policy");
     //CHECKSTYLE:ON
 
     @BeforeEach
     public void setUp() throws Exception {
         address = "http://" + opac.getHost() + ":" + opac.getMappedPort(opaPort);
         altAddress = "http://" + opac.getHost() + ":" + opac.getMappedPort(altPort) + "/customprefix";
-        eopaAddress = "http://" + eopac.getHost() + ":" + eopac.getMappedPort(opaPort);
-        eopaAltAddress = "http://" + eopac.getHost() + ":" + eopac.getMappedPort(altPort) + "/customprefix";
 
         this.recordingHandler = new RecordingHandler();
         this.recordingHandler.setLevel(ALL);
@@ -112,10 +93,6 @@ class OPATest {
         System.out.println("==== container logs from OPA container ====");
         final String logs = opac.getLogs();
         System.out.println(logs);
-
-        System.out.println("==== container logs from EOPA container ====");
-        final String eopalogs = eopac.getLogs();
-        System.out.println(eopalogs);
 
         this.logger.removeHandler(this.recordingHandler);
     }
@@ -145,29 +122,6 @@ class OPATest {
     }
 
     @Test
-    public void testEOPAHealth() {
-        // This test just makes sure that we can reach the OPAClient health endpoint
-        // and that it returns the expected {} value.
-
-        HttpClient client = HttpClient.newHttpClient();
-        HttpRequest req = HttpRequest.newBuilder().uri(URI.create(eopaAddress + "/health")).build();
-        HttpResponse<String> resp = null;
-
-        try {
-            resp = client.send(req, HttpResponse.BodyHandlers.ofString());
-        //CHECKSTYLE:OFF
-        } catch (Exception e) {
-            //CHECKSTYLE:ON
-            System.out.println("exception: " + e);
-            assertNull(e);
-        }
-
-        String body = resp.body();
-
-        assertEquals("{}\n", body);
-    }
-
-    @Test
     public void testOPAHealthAlternate() {
         // This makes sure that we can also successfully reach the OPA health
         // API on the "alternate", reverse-proxy based OPA that has a URL
@@ -180,30 +134,6 @@ class OPATest {
         try {
             resp = client.send(req, HttpResponse.BodyHandlers.ofString());
         // This is a unit test, I will catch whatever exceptions I want.
-        //CHECKSTYLE:OFF
-        } catch (Exception e) {
-            //CHECKSTYLE:ON
-            System.out.println("exception: " + e);
-            assertNull(e);
-        }
-
-        String body = resp.body();
-
-        assertEquals("{}\n", body);
-    }
-
-    @Test
-    public void testEOPAHealthAlternate() {
-        // This makes sure that we can also successfully reach the OPA health
-        // API on the "alternate", reverse-proxy based OPA that has a URL
-        // prefix.
-
-        HttpClient client = HttpClient.newHttpClient();
-        HttpRequest req = HttpRequest.newBuilder().uri(URI.create(eopaAltAddress + "/health")).build();
-        HttpResponse<String> resp = null;
-
-        try {
-            resp = client.send(req, HttpResponse.BodyHandlers.ofString());
         //CHECKSTYLE:OFF
         } catch (Exception e) {
             //CHECKSTYLE:ON
@@ -597,71 +527,6 @@ class OPATest {
         assertEquals(expect, result);
     }
 
-        @Test
-    public void testEOPAEvaluateBatch() {
-        OPAClient opa = new OPAClient(eopaAddress, headers);
-        Map<String, Object> input = Map.ofEntries(
-            entry("job1", Map.ofEntries(entry("aaa", "111"))),
-            entry("job2", Map.ofEntries(entry("bbb", "222")))
-        );
-        Map<String, OPAResult> result = Map.ofEntries();
-        Map<String, Object> expect = Map.ofEntries(
-            entry("job1", Map.ofEntries(entry("aaa", "111"))),
-            entry("job2", Map.ofEntries(entry("bbb", "222")))
-        );
-
-        try {
-            result = opa.evaluateBatch("policy/echo", input);
-        } catch (OPAException e) {
-            System.out.println("exception: " + e);
-            assertNull(e);
-        }
-
-        for (Map.Entry<String, OPAResult> entry: result.entrySet()) {
-            assertTrue(entry.getValue().success());
-            try {
-                Map<String, Object> v = entry.getValue().get();
-                assertEquals(expect.get(entry.getKey()), v);
-            } catch (OPAException e) {
-                System.out.println("exception: " + e);
-                assertNull(e);
-            }
-        }
-    }
-
-    @Test
-    public void testEOPAEvaluateBatchFallback() {
-        OPAClient opa = new OPAClient(eopaAddress, headers);
-        opa.forceBatchFallback(true);
-        Map<String, Object> input = Map.ofEntries(
-            entry("job1", Map.ofEntries(entry("aaa", "111"))),
-            entry("job2", Map.ofEntries(entry("bbb", "222")))
-        );
-        Map<String,  OPAResult> result = Map.ofEntries();
-        Map<String, Object> expect = Map.ofEntries(
-            entry("job1", Map.ofEntries(entry("aaa", "111"))),
-            entry("job2", Map.ofEntries(entry("bbb", "222")))
-        );
-
-        try {
-            result = opa.evaluateBatch("policy/echo", input);
-        } catch (OPAException e) {
-            System.out.println("exception: " + e);
-            assertNull(e);
-        }
-
-        for (Map.Entry<String, OPAResult> entry: result.entrySet()) {
-            assertTrue(entry.getValue().success());
-            try {
-                Map<String, Object> v = entry.getValue().get();
-                assertEquals(expect.get(entry.getKey()), v);
-            } catch (OPAException e) {
-                System.out.println("exception: " + e);
-                assertNull(e);
-            }
-        }
-    }
-
     @Test
     public void testOPAEvaluateBatchFallback() {
         OPAClient opa = new OPAClient(address, headers);
@@ -730,32 +595,5 @@ class OPATest {
             }
         }
     }
-
-    //@Test
-    //public void testEOPAEvaluateBatchMixed() {
-    //    OPAClient opa = new OPAClient(eopaAddress, headers);
-    //    Map<String, Object> input = Map.ofEntries(
-    //        entry("job1", Map.ofEntries(entry("aaa", "111"), entry("bbb", "111"))),
-    //        entry("job2", Map.ofEntries(entry("bbb", "222")))
-    //    );
-    //    Map<String, OPAResult> result = Map.ofEntries();
-    //    Map<String, Object> expect = Map.ofEntries(
-    //        entry("job1", Map.ofEntries(entry("aaa", "111"))),
-    //        entry("job2", Map.ofEntries(entry("bbb", "222")))
-    //    );
-    //
-    //    TypeReference<Map<String, Object>> tr = new TypeReference <Map<String, Object>>() {};
-    //
-    //    try {
-    //        result = opa.evaluateBatch("condfail/p", input, tr);
-    //    } catch (OPAException e) {
-    //        e.printStackTrace(System.out);
-    //        System.out.println("exception: " + e);
-    //        assertNull(e);
-    //    }
-    //
-    //    assertEquals(expect, result);
-    //}
-
 }
